+++ source/index.ts
@@ -20,6 +20,7 @@ import { DateTimeToSqlPipe } from "./pipes/formatting/date-time-to-sql";
 import { DateTimeToUtcPipe } from "./pipes/zones/date-time-to-utc";
 import { DurationFromIsoPipe } from "./pipes/parsing/duration-from-iso";
 import { DurationFromMillisecondsPipe } from "./pipes/parsing/duration-from-milliseconds";
+import { DurationToFormatPipe } from "./pipes/formatting/duration-to-format";
 import { DurationToIsoPipe } from "./pipes/formatting/duration-to-iso";
 
 const PIPES = [
@@ -43,6 +44,7 @@ const PIPES = [
   DateTimeToUtcPipe,
   DurationFromIsoPipe,
   DurationFromMillisecondsPipe,
+  DurationToFormatPipe,
   DurationToIsoPipe
 ];