+++ source/pipes/formatting/duration-to-format.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from "@angular/core";
+import { Duration } from "luxon";
+
+@Pipe({
+  name: "durationToFormat"
+})
+export class DurationToFormatPipe implements PipeTransform {
+
+  transform(value: Duration, format: string): string {
+    return value.toFormat(format);
+  }
+
+};