+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project are documented here. The format is based on [Keep a Changelog][keep-a-changelog].
 
+## Unreleased
+
+-   Added Duration -> arbitrary format pipe.
+
 ## 0.4.0 (2018-05-06)
 
 -   Aligned to Luxon ~1.2.0.